Replace the naive "always rescan" algorithm with a heap.
Add an implementation of heap.Interface to DataCollector so that
we can use container/heap to heapify the data.

Keep the values in a window like before, but also add two arrays:
 heap    - heap of pointers into the window
 revheap - pointers from the window into the heap

Since each addition of a sample adds modifies the heap, so each
addition of a sample is O(log m) m = heap size, so when we add
n samples, the total time complexity is O(n log m).